requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
Appium-Python-Client==1.2.0
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
execnet==1.9.0
importlib-metadata==4.5.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyHamcrest==2.0.2
pyparsing==2.4.7
pytest==6.2.4
pytest-async==0.1.1
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==10.0
pytest-xdist==2.3.0
PyYAML==5.4.1
selenium==3.141.0
Shelves==0.3.8
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
zipp==3.4.1
